requirements (1).txt

python-docx
pdfplumber
streamlit
